Move UploadError Kind to DirectUpload file
